refactor(email_api): extract 6 helpers purs vers helpers.rs
Extrait de main.rs (1556 LOC monstre) :
- normalize_segment (20 LOC) — normalisation Unicode
- build_misfits_local (9 LOC) — construction local misfits
- normalize_oauth_provider (30 LOC) — normalisation provider OAuth
- req_ip_str (10 LOC) — extraction IP client
- get_accept_language (8 LOC) — parsing Accept-Language
- welcome_email_html (59 LOC) — template HTML welcome

Ces fonctions sont sans dépendance métier, purement techniques.
Bénéfices : main.rs 1556 → 1417 LOC (−9%), helpers testables isolément.

Premier pas d'un split plus large de email_api_dir/main.rs. Les blocs
events/, mailing_list/, send_handler nécessitent d'exposer plus de
types partagés (MailEvent, EventBus, ComposerRecipient) — à faire dans
un PR ultérieur pour limiter le risque.
