+++ src/bin/email_api_dir/main.rs
@@ -54,6 +54,8 @@ mod mailbox;
 mod admin_ops;
 mod external_handlers;
 mod external_probe_handlers;
+mod helpers;
+use helpers::{normalize_segment, build_misfits_local, normalize_oauth_provider, req_ip_str, get_accept_language, welcome_email_html};
 
 pub use auth_handlers::*;
 pub use monitoring_handlers::*;
@@ -336,68 +338,6 @@ struct DeliverabilityReminderUpdate {
 // monitoring + security handlers → monitoring_handlers module
 
 
-fn normalize_segment(s: &str) -> String {
-    s.chars()
-        .map(|c| match c {
-            'à' | 'â' | 'ä' => 'a',
-            'é' | 'è' | 'ê' | 'ë' => 'e',
-            'î' | 'ï' => 'i',
-            'ô' | 'ö' => 'o',
-            'ù' | 'û' | 'ü' => 'u',
-            'ç' => 'c',
-            'ñ' => 'n',
-            'æ' => 'a',
-            'œ' => 'o',
-            _ => c,
-        })
-        .filter(|c| c.is_ascii_alphanumeric())
-        .map(|c| c.to_ascii_lowercase())
-        .collect()
-}
-
-/// Retourne `prenom.nom` normalisé, ou None si prénom ou nom est absent.
-fn build_misfits_local(first_name: &str, last_name: &str) -> Option<String> {
-    let first = normalize_segment(first_name.trim());
-    let last = normalize_segment(last_name.trim());
-    if first.is_empty() || last.is_empty() {
-        return None;
-    }
-    Some(format!("{}.{}", first, last))
-}
-
-fn normalize_oauth_provider(provider: &str) -> Option<String> {
-    let p = provider.trim().to_ascii_lowercase();
-    match p.as_str() {
-        "github" => Some(p),
-        _ => None,
-    }
-}
-
-#[derive(Deserialize, Serialize, Debug, PartialEq)]
-pub struct EmailRequest {
-    from: String,
-    to: String,
-    subject: String,
-    body: String,
-}
-
-#[derive(Deserialize)]
-struct MailingListRequest {
-    label: String,
-    emails: Vec<String>,
-}
-
-#[derive(Deserialize)]
-struct MailingListEmailRequest {
-    from: String,
-    subject: String,
-    body: String,
-    mailing_list: String,
-}
-
-async fn create_mailing_list(mailing_list: web::Json<MailingListRequest>) -> impl Responder {
-    let mailing_list_dir = Path::new("mailing-lists");
-    if !mailing_list_dir.exists() {
         if let Err(e) = create_dir_all(mailing_list_dir) {
             return HttpResponse::InternalServerError().json(serde_json::json!({
                 "status": "error",
@@ -584,86 +524,6 @@ async fn send_email_handler(
 }
 
 // --- Auth handlers ---
-
-fn req_ip_str(req: &actix_web::HttpRequest) -> String {
-    req.connection_info()
-        .realip_remote_addr()
-        .unwrap_or("unknown")
-        .split(':')
-        .next()
-        .unwrap_or("unknown")
-        .to_string()
-}
-
-fn get_accept_language(req: &actix_web::HttpRequest) -> String {
-    req.headers()
-        .get("accept-language")
-        .and_then(|v| v.to_str().ok())
-        .unwrap_or("")
-        .to_string()
-}
-
-fn welcome_email_html(
-    locale: &str,
-    display_name: &str,
-    primary_email: &str,
-    alias_email: Option<&str>,
-) -> String {
-    let dir = if i18n::is_rtl(locale) { "rtl" } else { "ltr" };
-    let text_align = if i18n::is_rtl(locale) {
-        "right"
-    } else {
-        "left"
-    };
-    let greeting = i18n::t(locale, "email-welcome-greeting", &[("name", display_name)]);
-    let intro = i18n::t(locale, "email-welcome-intro", &[]);
-    let primary_label = i18n::t(locale, "email-welcome-primary-label", &[]);
-    let cta = i18n::t(locale, "email-welcome-cta", &[]);
-    let signature = i18n::t(locale, "email-welcome-signature", &[]);
-    let alias_row = alias_email
-        .map(|a| {
-            let lbl = i18n::t(locale, "email-welcome-alias-label", &[]);
-            let detail = i18n::t(locale, "email-welcome-alias-detail", &[]);
-            format!(
-                "<tr><td style=\"padding:4px 0\"><b>{lbl}</b> \
-                 <span class=\"addr\">{a}</span> {detail}</td></tr>"
-            )
-        })
-        .unwrap_or_default();
-    format!(
-        r#"<!DOCTYPE html>
-<html lang="{locale}" dir="{dir}">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width,initial-scale=1.0">
-  <style>
-    body{{font-family:Tahoma,Arial,sans-serif;direction:{dir};text-align:{text_align};color:#222;background:#f5f5f5;margin:0;padding:0}}
-    .wrap{{max-width:600px;margin:40px auto;background:#fff;border-radius:8px;padding:40px}}
-    .addr{{font-family:monospace;background:#f0f0f0;padding:2px 6px;border-radius:3px}}
-  </style>
-</head>
-<body><div class="wrap">
-  <h2>{greeting}</h2>
-  <p>{intro}</p>
-  <table cellpadding="0" cellspacing="0">
-    <tr><td style="padding:4px 0"><b>{primary_label}</b> <span class="addr">{primary_email}</span></td></tr>
-    {alias_row}
-  </table>
-  <p>{cta}</p>
-  <p>{signature}</p>
-</div></body>
-</html>"#,
-    )
-}
-
-// auth_login..api_password_reset_confirm → auth_handlers module
-
-// mailbox + send + drafts → mailbox_handlers module
-// admin CRUD + change-requests + deliverability → admin_ops_handlers module
-
-#[actix_web::main]
-async fn main() -> std::io::Result<()> {
-    dotenv().ok();
 
     // rustls 0.23 requires an explicit process-level CryptoProvider.
     rustls::crypto::aws_lc_rs::default_provider()
